refactors the functions into their own files.
